requirements.txt
# smart-literature-agent 依赖（Python >= 3.10）
# 国内用户建议使用清华镜像：
#   pip install -r requirements.txt -i https://pypi.tuna.tsinghua.edu.cn/simple

# 核心：DeepXiv SDK（CLI + MCP，提供论文检索与全文解析）
deepxiv-sdk>=0.2.5

# 配置与调度
pyyaml>=6.0
schedule>=1.2.0

# LLM：Anthropic Python SDK（可直连官方 API，也可通过任何兼容 Messages API 的代理）
anthropic>=0.39.0

# 报告渲染：Markdown → HTML
markdown>=3.4
